+++ src/domain/event/eventPage.module.scss
@@ -22,6 +22,23 @@
   grid-template-columns: 1fr auto;
 }
 
+.cancelEnrolmentButton {
+  margin-top: 1rem;
+  margin-bottom: -1rem;
+}
+
+.enrolmentFormSection {
+  background-color: var(--enrolmentFormBackgroundColor);
+  padding: 0 var(--spacing-m);
+  border-top: 1px solid var(--color-black-10);
+  margin-top: -1px;
+
+  @include respond-above(sm) {
+    margin: 0 calc(-1 * var(--spacing-2-xl));
+    margin-top: -1px;
+  }
+}
+
 .eventHero {
   --back-button-padding: 1rem;
   --hero-height: 500px;
@@ -51,4 +68,14 @@
       height: 50px;
     }
   }
+}
+
+.occurrencesContainer {
+  @include respond-above(sm) {
+    margin: 0 calc(-1 * var(--spacing-2-xl));
+  }
+}
+
+.hideEnrolmentForm {
+  display: none;
 }
