+++ src/domain/venue/venueInfo/venueInfo.module.scss
@@ -1,7 +1,7 @@
 .venueInfo {
   --icon-size: 1.5rem;
   --icon-spacing: var(--spacing-xs);
-  --section-spacing: var(--spacing-l) 0;
+  --section-spacing: var(--spacing-l) 0 0 0;
 }
 
 .venueDescription {
